WIP: refactor mysql module to flatten its structure
