+++ quaint/src/connector/mysql.rs
@@ -1,8 +1,11 @@
-pub use wasm::common::MysqlUrl;
-pub use wasm::error::MysqlError;
+//! Wasm-compatible definitions for the MySQL connector.
+//! This module is only available with the `mysql` feature.
+pub mod error;
+pub(crate) mod url;
 
-#[cfg(feature = "mysql")]
-pub(crate) mod wasm;
+pub use error::MysqlError;
+pub use url::MysqlUrl;
 
 #[cfg(feature = "mysql-native")]
 pub(crate) mod native;
+
