+++ quaint/src/connector/mysql/error.rs
